requirements.txt
keras
tensorflow
pyserial
playsound
opencv-python
pillow
scikit-learn
seaborn
numpy
